+++ packages/api/api/model/common/__init__.py
@@ -1,1 +1,2 @@
 from .entity_name import EntityName, EntityNameData
+from .postal_address import PostalAddress, PostalAddressData
